+++ samples/sampler/src/ThemeEnvironment/ThemeEnvironment.js
@@ -41,7 +41,7 @@ const PanelsBase = kind({
 	)
 });
 
-const Theme = ThemeDecorator({overlay: false}, PanelsBase);
+const Theme = ThemeDecorator({overlay: false, rootId: 'storybook-root'}, PanelsBase);
 
 const StorybookDecorator = (story, config = {}) => {
 	// Executing `story` here allows the story controls to register and render before the global variable below.
@@ -64,6 +64,7 @@ const StorybookDecorator = (story, config = {}) => {
 	if (Object.keys(classes).length > 0) {
 		classes.debug = true;
 	}
+console.log(config.viewMode)
 
 	return (
 		<Theme
@@ -74,7 +75,8 @@ const StorybookDecorator = (story, config = {}) => {
 			textSize={JSON.parse(globals['large text']) ? 'large' : 'normal'}
 			highContrast={JSON.parse(globals['high contrast'])}
 			style={{
-				'--sand-env-background': globals.background === 'default' ? '' : globals.background
+				'--sand-env-background': globals.background === 'default' ? '' : globals.background,
+				//'--sand-docs-mode': config.viewMode === 'docs' ? 'static' : 'absolute'
 			}}
 			skin={globals.skin}
 			{...hasProps ? config.parameters.props : null}